+++ lightningd/test/run-invoice-select-inchan.c
@@ -34,12 +34,12 @@ void broadcast_tx(struct chain_topology *topo UNNEEDED,
 				 int exitstatus UNNEEDED,
 				 const char *err))
 { fprintf(stderr, "broadcast_tx called!\n"); abort(); }
-/* Generated stub for channel_tell_funding_locked */
-bool channel_tell_funding_locked(struct lightningd *ld UNNEEDED,
+/* Generated stub for channel_tell_depth */
+bool channel_tell_depth(struct lightningd *ld UNNEEDED,
 				 struct channel *channel UNNEEDED,
 				 const struct bitcoin_txid *txid UNNEEDED,
 				 u32 depth UNNEEDED)
-{ fprintf(stderr, "channel_tell_funding_locked called!\n"); abort(); }
+{ fprintf(stderr, "channel_tell_depth called!\n"); abort(); }
 /* Generated stub for command_fail */
 struct command_result *command_fail(struct command *cmd UNNEEDED, int code UNNEEDED,
 				    const char *fmt UNNEEDED, ...)
@@ -382,7 +382,7 @@ void set_log_outfn_(struct log_book *lr UNNEEDED,
 				  bool continued UNNEEDED,
 				  const struct timeabs *time UNNEEDED,
 				  const char *str UNNEEDED,
-				  const u8 *io UNNEEDED,
+				  const u8 *io UNNEEDED, size_t io_len UNNEEDED,
 				  void *arg) UNNEEDED,
 		    void *arg UNNEEDED)
 { fprintf(stderr, "set_log_outfn_ called!\n"); abort(); }