+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - JSON API: `newaddr` outputs `bech32` or `p2sh-segwit`, or both with new `all` parameter (#2390)
+- JSON API: `listpeers` status now shows how many confirmations until channel is open (#2405)
 
 ### Changed
 
@@ -24,6 +25,7 @@ changes.
 ### Fixed
 
 - `--bind-addr=<path>` fixed for nodes using local sockets (eg. testing).
+- Unannounced local channels were forgotten for routing on restart until reconnection occurred.
 
 ### Security